Simplify the use of @pytest.mark.integration()
